dev/requirements.txt
requests>=2.28.0
PyGithub>=1.58.0
gitpython>=3.1.30